+++ src/babel-plugin-fecool-helper/index.js
@@ -3,6 +3,8 @@ const { declare } = require("@babel/helper-plugin-utils");
 const { types } = require("@babel/core");
 const t = types;
 
+const { getOptions } = require("../config");
+
 module.exports = declare((api, options, dirname) => {
   api.assertVersion(7);
   const { loose, allowTopLevelThis, strict, strictMode, noInterop } = options;
@@ -10,15 +12,86 @@ module.exports = declare((api, options, dirname) => {
     name: "fecool-helper",
     visitor: {
       Program: {
-        enter(path, { opts }) {
+        enter(path, state) {
           const { parent } = path;
           const { comments = [] } = parent;
           if (!comments.length) return;
           const comment = comments[0];
           const { type, value } = comment;
-          if (type === "CommentBlock" && value && value.trim() === "@umdjs") {
-            path.node.sourceType = "script";
+          // 通过注释修改program的sourceType
+          if (
+            type === "CommentBlock" &&
+            value &&
+            /^@sourcetype=/.test(value.trim())
+          ) {
+            this.file.set(
+              "sourceType",
+              value.trim().replace(/^@sourcetype=/, "")
+            );
+            return;
+          } else if (
+            type === "CommentBlock" &&
+            value &&
+            value.trim() === "@tinytooljs"
+          ) {
+            this.file.set("sourceType", "tinytooljs");
+            return;
+          } else {
+            // 尝试访问获取当前解析的包类型
+            path.traverse({
+              ExpressionStatement: path => {
+                if (this.file.get("sourceType")) return;
+                if (!t.isProgram(path.parent)) return;
+                // 开始amd解析
+                if (
+                  t.isIdentifier(path.node.expression.callee, {
+                    name: "define"
+                  }) ||
+                  t.isIdentifier(path.node.expression.callee, {
+                    name: "require"
+                  })
+                ) {
+                  const args = path.node.expression.arguments || [];
+                  if (!args.length) return;
+                  const lastArgument = args[args.length - 1];
+                  if (t.isFunctionExpression(lastArgument)) {
+                    this.file.set("sourceType", "amd");
+                    path.stop();
+                  }
+                }
+                // 开始umd解析
+                if (
+                  t.isFunctionExpression(path.node.expression.callee) &&
+                  t.isBlockStatement(path.node.expression.callee.body)
+                ) {
+                  const args = path.node.expression.arguments || [];
+                  if (!args.length) return;
+                  const lastArgument = args[args.length - 1];
+                  if (t.isFunctionExpression(lastArgument)) {
+                    this.file.set("sourceType", "umd");
+                    path.stop();
+                  }
+                }
+              },
+              AssignmentExpression: path => {
+                // 开始commonjs解析
+                const left = path.get("left");
+                if (left && left.isMemberExpression()) {
+                  const object = left.get("object");
+                  if (object && object.node.name != "module") return;
+                  const property = left.get("property");
+                  if (property && property.node.name != "exports") return;
+                  this.file.set("sourceType", "commonjs");
+                  path.stop();
+                }
+              }
+            });
           }
+          const sourceType = this.file.get("sourceType"); // 'umd' 'amd' 'commonjs' 'tinytooljs'
+          path.node.sourceType =
+            sourceType && sourceType != "module"
+              ? "script"
+              : path.node.sourceType;
         }
       }
     }