+++ src/loader/json-loader/index.js
@@ -2,10 +2,10 @@ const path = require("path");
 const fs = require("fs");
 const { getOptions } = require("../../config");
 const { printer } = require("../../util");
-const { template } = require("@babel/core");
+const babel = require("@babel/core");
 const generate = require("@babel/generator").default;
 
-const codeWrapper = template(`
+const codeWrapper = babel.template(`
   define('NAME', function() {
     return JSON.parse('VALUE');
   })
@@ -50,9 +50,27 @@ module.exports = function({ dependName, paramName, filename }, options) {
         JSON: "JSON"
       });
       const code = generate(ast).code;
-      fs.writeFile(module.distFilename, code, "utf8", error => {
-        if (error) throw err;
-      });
+      if (getOptions().isDevelopENV()) {
+        fs.writeFile(module.distFilename, code, "utf8", error => {
+          if (error) throw err;
+        });
+      } else {
+        babel.transform(
+          code,
+          {
+            configFile: false,
+            babelrc: false,
+            minified: true,
+            compact: true
+          },
+          function(error, result) {
+            if (error) throw err;
+            fs.writeFile(module.distFilename, result.code, "utf8", error => {
+              if (error) throw err;
+            });
+          }
+        );
+      }
     }
   } catch (error) {
     printer.error(error);